First pass at writing capapbilties file
